+++ pkgs/ffigen/test/native_objc_test/bad_override_test.m
@@ -3,6 +3,37 @@
 // BSD-style license that can be found in the LICENSE file.
 
 #import <Foundation/NSObject.h>
+#import <Foundation/NSString.h>
+
+@interface Polygon : NSObject {}
+-(NSString*)name;
+@end
+@implementation Polygon
+-(NSString*)name { return @"Polygon"; }
+@end
+
+@interface Triangle : Polygon {}
+-(NSString*)name;
+@end
+@implementation Triangle
+-(NSString*)name { return @"Triangle"; }
+@end
+
+@interface Rectangle : Polygon {}
+-(NSString*)name;
+@end
+@implementation Rectangle
+-(NSString*)name { return @"Rectangle"; }
+@end
+
+@interface Square : Rectangle {}
+-(NSString*)name;
+@end
+@implementation Square
+-(NSString*)name { return @"Square"; }
+@end
+
+
 
 @interface BadOverrideGrandparent : NSObject {}
 @end
@@ -11,10 +42,16 @@
 
 @interface BadOverrideParent : BadOverrideGrandparent {}
 -(int32_t)methodVsGetter;
-@property (readonly) int32_t methodVsGetter;
+-(Rectangle*)contravariantReturn;
+-(NSString*)covariantArg:(Polygon*)poly;
 @end
 @implementation BadOverrideParent
 -(int32_t)methodVsGetter { return 1; }
+-(Rectangle*)contravariantReturn { return [Rectangle new]; }
+
+-(NSString*)covariantArg:(Polygon*)poly {
+  return [@"Polygon: " stringByAppendingString: [poly name]];
+}
 @end
 
 @interface BadOverrideUncle : BadOverrideGrandparent {}
@@ -31,9 +68,16 @@
 
 @interface BadOverrideChild : BadOverrideParent {}
 @property (readonly) int32_t methodVsGetter;
+-(Polygon*)contravariantReturn;
+-(NSString*)covariantArg:(Rectangle*)rect;
 @end
 @implementation BadOverrideChild
 -(int32_t)methodVsGetter { return 11; }
+-(Polygon*)contravariantReturn { return [Triangle new]; }
+
+-(NSString*)covariantArg:(Rectangle*)rect {
+  return [@"Rectangle: " stringByAppendingString: [rect name]];
+}
 @end
 
 @interface BadOverrideSibbling : BadOverrideParent {}